PYI-19: Refactoring and cleaning up.
