+++ src/main/java/uk/gov/di/ipv/core/back/service/AttributeCollectionService.java
@@ -1,6 +1,7 @@
 package uk.gov.di.ipv.core.back.service;
 
 import uk.gov.di.ipv.core.back.domain.AttributeName;
+import uk.gov.di.ipv.core.back.domain.SessionData;
 import uk.gov.di.ipv.core.back.domain.data.IdentityEvidence;
 
 import java.util.List;
@@ -9,4 +10,6 @@ import java.util.Map;
 public interface AttributeCollectionService {
 
     Map<AttributeName, List<String>> collectAttributesFromEvidence(IdentityEvidence identityEvidence);
+
+    void updateAttributesInSession(SessionData sessionData, IdentityEvidence identityEvidence);
 }
